+++ webapi/src/main/scala/org/knora/webapi/messages/util/search/QueryTraverser.scala
@@ -15,6 +15,10 @@ import org.knora.webapi.messages.SmartIri
 import org.knora.webapi.messages.StringFormatter
 import org.knora.webapi.messages.util.search.gravsearch.prequery.AbstractPrequeryGenerator
 import org.knora.webapi.slice.ontology.repo.service.OntologyCache
+import org.knora.webapi.messages.util.search.gravsearch.transformers.{
+  SelectToSelectTransformer,
+  ConstructToConstructTransformer
+}
 
 /**
  * A trait for classes that visit statements and filters in WHERE clauses, accumulating some result.
