+++ webapi/src/main/scala/org/knora/webapi/messages/util/search/gravsearch/transformers/SparqlTransformer.scala
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.knora.webapi.messages.util.search
+package org.knora.webapi.messages.util.search.gravsearch.transformers
 
 import zio.ZLayer
 import zio._
@@ -15,6 +15,7 @@ import org.knora.webapi.messages.IriConversions._
 import org.knora.webapi.messages.OntologyConstants
 import org.knora.webapi.messages.SmartIri
 import org.knora.webapi.messages.StringFormatter
+import org.knora.webapi.messages.util.search._
 import org.knora.webapi.slice.ontology.repo.service.OntologyCache
 
 import SparqlTransformer._
